frame:  - removed compiler warnings.
about:
 - added project link;

dateclock:
 - added a mouse click event to open the popup window when clicked;
 - removed defs.h.

popup: 
 - created a simple popup class.

git-svn-id: https://rwm-project.googlecode.com/svn/trunk@21 dac65600-f84a-dd98-4646-59ecd36c4ea3

+++ src/frame.cpp
@@ -232,8 +232,8 @@ void Frame::get_wm_hints()
 void Frame::get_wm_normal_hints() // Poor implementation of many applications ....
 {
     XSizeHints *xsizehints;
-    long hints = NULL;
-    long wm_flags = NULL;
+    long hints = (long)NULL;
+    long wm_flags = (long)NULL;
 
     if ((xsizehints = XAllocSizeHints()) != NULL)
     {
